YQL-19747 Extract names module
