+++ yql/essentials/sql/v1/complete/name/ya.make
@@ -1,8 +1,12 @@
 LIBRARY()
 
+SRCS(
+    name.cpp
+    parse.cpp
+)
+
 PEERDIR(
     yql/essentials/sql/v1/complete/core
-    yql/essentials/sql/v1/complete/name/service
 )
 
 END()
